requirements.txt
aws-cdk.core==1.104.0
aws-cdk.aws_events==1.104.0
aws-cdk.aws_lambda==1.104.0
aws-cdk.aws_events_targets==1.104.0
aws_cdk.aws_cloudwatch==1.104.0
aws_cdk.aws_sqs==1.104.0
aws_cdk.aws_lambda_event_sources==1.104.0
aws_cdk.aws_kms==1.104.0
